Restructure modules
- Move evaluators into new module evaluator.py.
- Rename cache.py to decorate.py
