+++ src/flask_cachecontrol/__init__.py
@@ -18,8 +18,8 @@
     :license: BSD, see LICENSE for more details.
 """
 
-from .cache import cache, cache_for, dont_cache, Always, ResponseIsSuccessful, \
-    ResponseIsSuccessfulOrRedirect
+from .decorate import cache, cache_for, dont_cache
+from .evaluator import Always, ResponseIsSuccessful, ResponseIsSuccessfulOrRedirect
 from .error import FlaskCacheControlError, CacheControlAttributeInvalidError
 
 __version__ = '0.2.1'
